status webhook

+++ Erbium/Engine/Private/NetDriver.cpp
@@ -512,8 +512,6 @@ void UNetDriver::TickFlush(UNetDriver* Driver, float DeltaSeconds)
 		static auto bSkipAircraft = GameState->CurrentPlaylistInfo.BasePlaylist ? GameState->CurrentPlaylistInfo.BasePlaylist->bSkipAircraft : false;
         if (!bSkipAircraft && GameState->HasWarmupCountdownEndTime() && Driver->ClientConnections.Num() > 0 && GameMode->bWorldIsReady && GameState->WarmupCountdownEndTime <= Time)
         {
-			GUI::gsStatus = 2;
-
             UKismetSystemLibrary::ExecuteConsoleCommand(UWorld::GetWorld(), FString(L"startaircraft"), nullptr);
         }
 	}
@@ -521,7 +519,39 @@ void UNetDriver::TickFlush(UNetDriver* Driver, float DeltaSeconds)
 	{
 		auto WorldNetDriver = UWorld::GetWorld()->NetDriver;
 		if (Driver == WorldNetDriver && Driver->ClientConnections.Num() == 0)
-			TerminateProcess(GetCurrentProcess(), 0);
+		{
+			static bool stopped = false;
+
+			if (!stopped)
+			{
+				stopped = true;
+
+				if constexpr (FConfiguration::WebhookURL && *FConfiguration::WebhookURL)
+				{
+					auto curl = curl_easy_init();
+
+					curl_easy_setopt(curl, CURLOPT_URL, FConfiguration::WebhookURL);
+					curl_slist* headers = curl_slist_append(NULL, "Content-Type: application/json");
+					curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
+
+					char version[6];
+
+					sprintf_s(version, VersionInfo.FortniteVersion >= 5.00 || VersionInfo.FortniteVersion < 1.2 ? "%.2f" : "%.1f", VersionInfo.FortniteVersion);
+
+					auto Playlist = FindObject<UFortPlaylistAthena>(FConfiguration::Playlist);
+					auto payload = UEAllocatedString("{\"embeds\": [{\"title\": \"Match has ended!\", \"fields\": [{\"name\":\"Version\",\"value\":\"") + version + "\"}, {\"name\":\"Playlist\",\"value\":\"" + (Playlist ? Playlist->PlaylistName.ToString() : "Playlist_DefaultSolo") + "\"}], \"color\": " + "\"7237230\", \"footer\": {\"text\":\"Erbium\", \"icon_url\":\"https://cdn.discordapp.com/attachments/1341168629378584698/1436803905119064105/L0WnFa.png.png?ex=6910ef69&is=690f9de9&hm=01a0888b46647959b38ee58df322048ab49e2a5a678e52d4502d9c5e3978d805&\"}, \"timestamp\":\"" + iso8601() + "\"}] }";
+
+					curl_easy_setopt(curl, CURLOPT_POSTFIELDS, payload.c_str());
+
+					curl_easy_perform(curl);
+
+					curl_easy_cleanup(curl);
+				}
+
+				if (FConfiguration::bAutoRestart)
+					TerminateProcess(GetCurrentProcess(), 0);
+			}
+		}
 	}
 
     return TickFlushOG(Driver, DeltaSeconds);
@@ -546,15 +576,46 @@ void UNetDriver::TickFlush__RepGraph(UNetDriver* Driver, float DeltaSeconds)
 			static auto bSkipAircraft = GameState->CurrentPlaylistInfo.BasePlaylist ? GameState->CurrentPlaylistInfo.BasePlaylist->bSkipAircraft : false;
 			if (!bSkipAircraft && Driver->ClientConnections.Num() > 0 && GameMode->bWorldIsReady && GameState->WarmupCountdownEndTime <= Time)
 			{
-				GUI::gsStatus = 2;
-
 				UKismetSystemLibrary::ExecuteConsoleCommand(UWorld::GetWorld(), FString(L"startaircraft"), nullptr);
 			}
 		}
-		else if (GUI::gsStatus == 2 && FConfiguration::bAutoRestart)
+		else if (GUI::gsStatus == 2)
 		{
-			if (Driver->ClientConnections.Num() == 0)
-				TerminateProcess(GetCurrentProcess(), 0);
+			auto WorldNetDriver = UWorld::GetWorld()->NetDriver;
+			if (Driver == WorldNetDriver && Driver->ClientConnections.Num() == 0)
+			{
+				static bool stopped = false;
+
+				if (!stopped)
+				{
+					stopped = true;
+
+					if constexpr (FConfiguration::WebhookURL && *FConfiguration::WebhookURL)
+					{
+						auto curl = curl_easy_init();
+
+						curl_easy_setopt(curl, CURLOPT_URL, FConfiguration::WebhookURL);
+						curl_slist* headers = curl_slist_append(NULL, "Content-Type: application/json");
+						curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
+
+						char version[6];
+
+						sprintf_s(version, VersionInfo.FortniteVersion >= 5.00 || VersionInfo.FortniteVersion < 1.2 ? "%.2f" : "%.1f", VersionInfo.FortniteVersion);
+
+						auto Playlist = FindObject<UFortPlaylistAthena>(FConfiguration::Playlist);
+						auto payload = UEAllocatedString("{\"embeds\": [{\"title\": \"Match has ended!\", \"fields\": [{\"name\":\"Version\",\"value\":\"") + version + "\"}, {\"name\":\"Playlist\",\"value\":\"" + (Playlist ? Playlist->PlaylistName.ToString() : "Playlist_DefaultSolo") + "\"}], \"color\": " + "\"7237230\", \"footer\": {\"text\":\"Erbium\", \"icon_url\":\"https://cdn.discordapp.com/attachments/1341168629378584698/1436803905119064105/L0WnFa.png.png?ex=6910ef69&is=690f9de9&hm=01a0888b46647959b38ee58df322048ab49e2a5a678e52d4502d9c5e3978d805&\"}, \"timestamp\":\"" + iso8601() + "\"}] }";
+
+						curl_easy_setopt(curl, CURLOPT_POSTFIELDS, payload.c_str());
+
+						curl_easy_perform(curl);
+
+						curl_easy_cleanup(curl);
+					}
+
+					if (FConfiguration::bAutoRestart)
+						TerminateProcess(GetCurrentProcess(), 0);
+				}
+			}
 		}
 	}
 
@@ -626,7 +687,39 @@ void UNetDriver::TickFlush__Iris(UNetDriver* Driver, float DeltaSeconds)
 	{
 		auto WorldNetDriver = UWorld::GetWorld()->NetDriver;
 		if (Driver == WorldNetDriver && Driver->ClientConnections.Num() == 0)
-			TerminateProcess(GetCurrentProcess(), 0);
+		{
+			static bool stopped = false;
+
+			if (!stopped)
+			{
+				stopped = true;
+
+				if constexpr (FConfiguration::WebhookURL && *FConfiguration::WebhookURL)
+				{
+					auto curl = curl_easy_init();
+
+					curl_easy_setopt(curl, CURLOPT_URL, FConfiguration::WebhookURL);
+					curl_slist* headers = curl_slist_append(NULL, "Content-Type: application/json");
+					curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
+
+					char version[6];
+
+					sprintf_s(version, VersionInfo.FortniteVersion >= 5.00 || VersionInfo.FortniteVersion < 1.2 ? "%.2f" : "%.1f", VersionInfo.FortniteVersion);
+
+					auto Playlist = FindObject<UFortPlaylistAthena>(FConfiguration::Playlist);
+					auto payload = UEAllocatedString("{\"embeds\": [{\"title\": \"Match has ended!\", \"fields\": [{\"name\":\"Version\",\"value\":\"") + version + "\"}, {\"name\":\"Playlist\",\"value\":\"" + (Playlist ? Playlist->PlaylistName.ToString() : "Playlist_DefaultSolo") + "\"}], \"color\": " + "\"7237230\", \"footer\": {\"text\":\"Erbium\", \"icon_url\":\"https://cdn.discordapp.com/attachments/1341168629378584698/1436803905119064105/L0WnFa.png.png?ex=6910ef69&is=690f9de9&hm=01a0888b46647959b38ee58df322048ab49e2a5a678e52d4502d9c5e3978d805&\"}, \"timestamp\":\"" + iso8601() + "\"}] }";
+
+					curl_easy_setopt(curl, CURLOPT_POSTFIELDS, payload.c_str());
+
+					curl_easy_perform(curl);
+
+					curl_easy_cleanup(curl);
+				}
+
+				if (FConfiguration::bAutoRestart)
+					TerminateProcess(GetCurrentProcess(), 0);
+			}
+		}
 	}
 
 	return TickFlushOG(Driver, DeltaSeconds);
